backend/pyproject.toml
[project]
name = "deep-research-agent-backend"
version = "0.1.0"
description = "Backend-first hybrid architecture for Open Deep Research Agent"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "redis>=5.0.0",
  "motor>=3.6.0",
  "qdrant-client>=1.11.0",
  "langgraph>=0.2.35",
  "langchain-google-genai>=2.0.0",
  "tavily-python>=0.5.0",
  "sentence-transformers>=3.1.0",
  "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "httpx>=0.27.0",
  "anyio>=4.4.0"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["app*"]
